requirements.txt
asgiref==3.5.2
boto3==1.24.62
botocore==1.27.62
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
django-shortcuts==1.6
djangorestframework==3.13.1
idna==3.3
jmespath==1.0.1
joblib==1.1.0
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
pandas==1.4.3
psycopg2==2.9.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.64.0
tweepy==3.10.0
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.2.0
